+++ core-bundle/src/Resources/contao/library/Contao/InsertTags.php
@@ -725,7 +725,7 @@ class InsertTags extends \Controller
 						$elements[1] = 'mainTitle';
 					}
 
-					// Do not use \StringUtil::specialchars() here (see #4687)
+					// Do not use StringUtil::specialchars() here (see #4687)
 					$arrCache[$strTag] = $objPage->{$elements[1]};
 					break;
 
@@ -999,10 +999,7 @@ class InsertTags extends \Controller
 						}
 					}
 
-					$container
-						->get('monolog.logger.contao')
-						->log(LogLevel::INFO, 'Unknown insert tag: ' . $strTag)
-					;
+					$this->log('Unknown insert tag {{' . $strTag . '}}', __METHOD__, TL_ERROR);
 					break;
 			}
 
@@ -1093,10 +1090,7 @@ class InsertTags extends \Controller
 								}
 							}
 
-							$container
-								->get('monolog.logger.contao')
-								->log(LogLevel::INFO, 'Unknown insert tag flag: ' . $flag)
-							;
+							$this->log('Unknown insert tag flag "' . $flag . '" in {{' . $strTag . '}}', __METHOD__, TL_ERROR);
 							break;
 					}
 				}